+++ src/bindings/vips-operators.cpp
@@ -852,10 +852,9 @@ emscripten::val Image::profile_load(const std::string &name)
                     ->set("profile", &profile)
                     ->set("name", name));
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(profile)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(profile)->data)));
-    VIPS_AREA(profile)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(profile));
 
     return result;
@@ -2036,10 +2035,9 @@ emscripten::val Image::dzsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -2278,10 +2276,9 @@ emscripten::val Image::gifsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -2359,10 +2356,9 @@ emscripten::val Image::heifsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -2672,10 +2668,9 @@ emscripten::val Image::jp2ksave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -2709,10 +2704,9 @@ emscripten::val Image::jpegsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -2754,10 +2748,9 @@ emscripten::val Image::jxlsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -2832,10 +2825,9 @@ emscripten::val Image::magicksave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -3137,10 +3129,9 @@ emscripten::val Image::pngsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -3257,10 +3248,9 @@ emscripten::val Image::radsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -3765,10 +3755,9 @@ emscripten::val Image::tiffsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;
@@ -3850,10 +3839,9 @@ emscripten::val Image::webpsave_buffer(emscripten::val js_options) const
                    ->set("buffer", &buffer),
                js_options);
 
-    emscripten::val result = emscripten::val(emscripten::typed_memory_view(
+    emscripten::val result = BlobVal.new_(emscripten::typed_memory_view(
         VIPS_AREA(buffer)->length,
         reinterpret_cast<uint8_t *>(VIPS_AREA(buffer)->data)));
-    VIPS_AREA(buffer)->free_fn = nullptr;
     vips_area_unref(VIPS_AREA(buffer));
 
     return result;